+++ src/app/(payload)/admin/importMap.js
@@ -1,8 +1,10 @@
 import { AIProductAssistant as AIProductAssistant_70aca31e0bbb8a8f74cb195cc74eca10 } from '../../../admin/AIProductAssistant'
+import { AICategorySeoAssistant as AICategorySeoAssistant_e7c09871763c1409796269b7e4c486cc } from '../../../admin/AICategorySeoAssistant'
 import { CollectionCards as CollectionCards_f9c02e79a4aed9a3924487c0cd4cafb1 } from '@payloadcms/next/rsc'
 
 /** @type import('payload').ImportMap */
 export const importMap = {
   "/admin/AIProductAssistant#AIProductAssistant": AIProductAssistant_70aca31e0bbb8a8f74cb195cc74eca10,
+  "/admin/AICategorySeoAssistant#AICategorySeoAssistant": AICategorySeoAssistant_e7c09871763c1409796269b7e4c486cc,
   "@payloadcms/next/rsc#CollectionCards": CollectionCards_f9c02e79a4aed9a3924487c0cd4cafb1
 }